crypt: allow keyslots to be named
This updates the APIs in crypt.go that add / modify keys to create
named keyslots - a keyslot with an associated token with the type
"ubuntu-fde" or "ubuntu-fde-recovery" - each of these tokens has
a unique name.

- IntializeLUKS2Container now creates a named keyslot with the
type "ubuntu-fde" and the name "default" unless another name is
supplied.
- AddRecoveryKeyToLUKS2Container has been renamed to
SetLUKS2ContainerRecoveryKey and creates a named recovery keyslot
with the type "ubuntu-fde-recovery" and the name "default-recovery"
unless another name is supplied.
- ChangeLUKS2ContainerUsingRecoveryKey has been renamed to
SetLUKS2ContainerUnlockKey and creates a named keyslot with the type
"ubuntu-fde" and the name "default" unless another name is supplied.
- DeleteLUKS2ContainerKey deletes a named keyslot.
- ListLUKS2ContainerUnlockKeyNames and ListLUKS2ContainerRecoveryKeyNames
can be used to list the names of existing named keyslots.

A subsequent PR will provide an API for updating the "ubuntu-fde"
tokens in order to attach and update the associated KeyData.

The SetLUKS2Container*Key functions are not atomic and they cannot
be used to change the key associated with an existing keyslot. If
the name of an existing keyslot is supplied then it will be deleted
first (as long as it is not the last keyslot) and then recreated.
The ability to change keys isn't particularly useful anyway, because
although cryptsetup can change a key atomically, changing a
"ubuntu-fde" keyslot always requires 2 operations (change the keyslot
and then import the updated KeyData).